gemini_and_textract/gemini_requirements.txt
google-genai
python-dotenv
pdf2image
pydantic==2.10.6
matplotlib
pandas
beautifulsoup4